Moves alpha_plots into per sample
Because pool and pickle don’t play nicely with functions
